+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to the "js-console" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [v0.1.0]
+## [v0.6.0]
 
-initial release
+Enabled to color the text of console.log.
+
+settings < extensions < JavaScript Insert Console < js-console.textColor
+
+example
+
+```
+console.log('%c name: ', 'color: #112233', name);
+```
 
 ## [v0.5.0]
 
 ❌ console.table('variable: ', variable);
 ⭕️ console.table(variable);
+
+## [v0.1.0]
+
+initial release